[refactor]タイポ修正とvue名をUCCに変更
