+++ server/front/src/router.ts
@@ -3,10 +3,10 @@ import Router from 'vue-router';
 import Index from './views/Index.vue';
 import Login from './views/Login.vue';
 import Timeline from './views/Timeline.vue';
-import UserInitProfile from './views/User_Init_Profile.vue';
+import UserInitProfile from './views/UserInitProfile.vue';
 import UserPage from './views/UserPage.vue';
-import FollowList from './views/Follow_List.vue';
-import EditProfile from './views/Edit_Profile.vue';
+import FollowList from './views/FollowList.vue';
+import EditProfile from './views/EditProfile.vue';
 
 Vue.use(Router);
 
@@ -33,7 +33,7 @@ export default new Router({
 
     {
       path: '/user_init_profile',
-      name: 'User_init_profile',
+      name: 'UserInitProfile',
       component: UserInitProfile,
     },
