+++ server/front/src/views/UserInitProfile.vue
